Refactor EntityInfoPanel
